Fix the collisions to work with the client / server split
We were only calling `regenerateShapes()` on the client. Which meant `getCollisionShape()` always returned a full Cube on the server.

Thus, the client-side prediction would temporarily allow the player to go forward, but then the server would snap them back, because it thinks the player is cheating. Hence, the collision was choppy.

Now that we properly call `regenerateShapes()` on the server, they both match, and the collision is smooth